implement priority for groups

+++ simdb/databaseAPI.py
@@ -38,8 +38,10 @@ from sqlalchemy import or_, and_
 # define string types
 string_types = str if sys.version_info[0] == 3 else basestring
 
-
-Session = sessionmaker()
+session_info = {
+    "SIMDB_PRIORITY" : 0
+}
+Session = sessionmaker(info=session_info)
 
 
 def create_new_database(db_path, priority=None):
@@ -68,7 +70,7 @@ def create_new_database(db_path, priority=None):
     session = Session()
     Base.metadata.create_all(engine)
 
-    session.info["SIMDB_PRIORITY"] = priority or 0
+    session.info["SIMDB_PRIORITY"] = priority or session.info["SIMDB_PRIORITY"]
 
     return session
 
@@ -99,7 +101,7 @@ def connect_database(db_path, priority=None):
     Session.configure(bind=engine)
     session = Session()
 
-    session.info["SIMDB_PRIORITY"] = priority or 0
+    session.info["SIMDB_PRIORITY"] = priority or session.info["SIMDB_PRIORITY"]
 
     return session
 
@@ -131,7 +133,7 @@ def session_handler(db_path, create=False, priority=None):
         engine = create_engine('sqlite:///{}'.format(db_path))
         Session.configure(bind=engine)
         session = Session()
-        session.info["SIMDB_PRIORITY"] = priority or 0
+        session.info["SIMDB_PRIORITY"] = priority or session.info["SIMDB_PRIORITY"]
     else:
         raise IOError("No such file or directory: '{}'".format(db_path))
     try:
@@ -491,6 +493,7 @@ def add_single_keyword(session, entry_id, name, value=None, unique=True, main_id
         keyword = Keywords(name=name, value=value, main_id=main_id, priority=priority)
     else:
         if priority >= keyword.priority:
+            keyword.priority = priority
             keyword.value = value
 
     session.add(keyword)
@@ -518,6 +521,7 @@ def update_keywords(session, entry_id, priority=None, **kwargs):
         .filter(Keywords.name.in_(kwargs.keys())).all()
     for keyword in keywords:
         if priority >= keyword.priority:
+            keyword.priority = priority
             keyword.value = kwargs.pop(keyword.name)
 
     if len(keywords) != 0:
@@ -688,7 +692,7 @@ def get_single_meta_group(session, entry_id, name, as_list=False):
         return None
     return metagroup.to_list() if as_list else metagroup.to_dict()
 
-def add_meta_group(session, entry_id, meta_group_name, unique=True, main_id=None):
+def add_meta_group(session, entry_id, meta_group_name, unique=True, main_id=None, priority=None):
     """
     Add a meta information group/block to entry.
 
@@ -711,6 +715,9 @@ def add_meta_group(session, entry_id, meta_group_name, unique=True, main_id=None
         added MetaGroups object
     """
 
+    # evaluate priority
+    priority = priority or session.info["SIMDB_PRIORITY"]
+
     # get meta_group if unique else None
     meta_group = session.query(MetaGroups).join(Main)\
         .filter(MetaGroups.name == meta_group_name)\
@@ -721,13 +728,16 @@ def add_meta_group(session, entry_id, meta_group_name, unique=True, main_id=None
     if meta_group is None:
         if main_id is None:
             main_id = session.query(Main.id).filter(Main.entry_id == entry_id).one()[0]
-        meta_group = MetaGroups(name=meta_group_name, main_id=main_id)
+        meta_group = MetaGroups(name=meta_group_name, main_id=main_id, priority=priority)
         session.add(meta_group)
+    else:
+        meta_group.priority = max(priority, meta_group.priority)
+
 
     return meta_group
 
 
-def delete_meta_group(session, entry_id, meta_group_name):
+def delete_meta_group(session, entry_id, meta_group_name, priority=None):
     """
     Remove a meta information group/block from entry.
 
@@ -741,15 +751,19 @@ def delete_meta_group(session, entry_id, meta_group_name):
         Name of meta group, e.g. Thermostat, Barostat
     """
 
+    # evaluate priority
+    priority = priority or session.info["SIMDB_PRIORITY"]
+
     meta_groups = session.query(MetaGroups).join(Main) \
         .filter(MetaGroups.name == meta_group_name)\
         .filter(Main.entry_id == entry_id) \
         .all()
     for meta_group in meta_groups:
-        session.delete(meta_group)
+        if priority >= meta_group.priority:
+            session.delete(meta_group)
 
 
-def add_meta_data(session, entry_id, meta_group_name, unique=True, metagroup_id=None, **kwargs):
+def add_meta_data(session, entry_id, meta_group_name, unique=True, metagroup_id=None, priority=None, **kwargs):
     """
     Add meta data to entry.
 
@@ -775,6 +789,10 @@ def add_meta_data(session, entry_id, meta_group_name, unique=True, metagroup_id=
     meta_entries : List[MetaEntry]
         List of added MetaEntries
     """
+
+    # evaluate priority
+    priority = priority or session.info["SIMDB_PRIORITY"]
+
     meta_entries = session.query(MetaEntry).join(MetaGroups) \
         .filter(MetaGroups.name == meta_group_name)\
         .filter(MetaEntry.metagroup_id == MetaGroups.id)\
@@ -797,12 +815,12 @@ def add_meta_data(session, entry_id, meta_group_name, unique=True, metagroup_id=
                 metagroup_id = metagroup_id[0]
             else:
                 main_id = session.query(Main.id).filter(Main.entry_id == entry_id).one()[0]
-                meta_group = MetaGroups(name=meta_group_name, main_id=main_id)
+                meta_group = MetaGroups(name=meta_group_name, main_id=main_id, priority=priority)
                 session.add(meta_group)
                 session.flush()
                 metagroup_id = meta_group.id
     for name, value in kwargs.items():
-        meta_entries.append(MetaEntry(name=name, value=value, metagroup_id=metagroup_id))
+        meta_entries.append(MetaEntry(name=name, value=value, metagroup_id=metagroup_id, priority=priority))
     session.add_all(meta_entries)
 
     return meta_entries
@@ -927,10 +945,79 @@ def remove_meta_data(session, entry_id, meta_group_name, **kwargs):
 
 
 # =========================================================================== #
-# manage groups
+# MANAGE GROUPS                                                               #
 # =========================================================================== #
 
-def add_to_group(session, entry_id, group_name):
+def add_group(session, group_name, priority=None):
+    """Add group to database.
+
+    Parameters
+    ----------
+    session : sqlalchemy.orm.session.Session
+        SQL Alchemy session
+    group_name: str
+        Name of group
+    priority: int or None
+        use session priority if None
+    """
+
+    priority = priority if priority is not None else session.info["SIMDB_PRIORITY"]
+    group = session.query(Groups).filter(Groups.name == group_name).first()
+
+    if not group:
+        group = Groups(name=group_name, priority=priority)
+        session.add(group)
+    else:
+        if priority >= group.priority:
+            group.priority = priority
+
+
+def remove_group(session, group_name, priority=None):
+    """Remove group from database.
+
+    Parameters
+    ----------
+    session : sqlalchemy.orm.session.Session
+        SQL Alchemy session
+    group_name: str
+        Name of group
+    priority: int or None
+        use session priority if None
+    """
+
+    priority = priority if priority is not None else session.info["SIMDB_PRIORITY"]
+    group = session.query(Groups).filter(Groups.name == group_name).first()
+
+    if group:
+        if priority >= group.priority:
+            session.delete(group)
+
+
+def rename_group(session, group_name, new_group_name, priority=None):
+    """Rename group in database.
+
+    Parameters
+    ----------
+    session : sqlalchemy.orm.session.Session
+        SQL Alchemy session
+    group_name: str
+        Name of group
+    new_group_name: str
+        Name of group
+    priority: int or None
+        use session priority if None
+    """
+
+    priority = priority if priority is not None else session.info["SIMDB_PRIORITY"]
+    group = session.query(Groups).filter(Groups.name == group_name).first()
+
+    if group:
+        if priority >= group.priority:
+            group.name = new_group_name
+            group.priority = priority
+
+
+def add_to_group(session, entry_id, group_name, priority=None):
     """Add simulation to group.
 
     Parameters
@@ -941,12 +1028,10 @@ def add_to_group(session, entry_id, group_name):
         Entry ID in database
     group_name: str
         Name of group
-
-    Returns
-    -------
-    group : Groups
+    priority: int or None
+        use session priority if None
     """
-
+    priority = priority if priority is not None else session.info["SIMDB_PRIORITY"]
     entry = session.query(Main).filter(Main.entry_id == entry_id).first()
 
     if entry:
@@ -954,14 +1039,12 @@ def add_to_group(session, entry_id, group_name):
         # get group if already in DB or create new group
         group = session.query(Groups).filter(Groups.name == group_name).first()
         if not group:
-            group = Groups(name=group_name)
+            group = Groups(name=group_name, priority=priority)
             session.add(group)
-            session.flush()
 
         if group not in entry.groups:
-            group.entries.append(entry)
+            entry.groups.append(group)
 
-    return group
 
 
 def remove_from_group(session, entry_id, group_name):
@@ -975,10 +1058,6 @@ def remove_from_group(session, entry_id, group_name):
         Entry ID in database
     group_name: str
         Name of group
-
-    Returns
-    -------
-    True if entry was removed from group, otherwise False.
     """
 
     entry = session.query(Main).filter(Main.entry_id == entry_id).first()
@@ -1016,33 +1095,7 @@ def get_group_keywords(session, group_name):
         raise NoResultFound("group_name = '{}' was not found in database.".format(group_name))
 
 
-def set_group_keywords(session, group_name, **kwargs):
-    """
-    Function to set group keywords. Overwrites the old!
-
-    Parameters
-    ----------
-    session : sqlalchemy.orm.session.Session
-        SQL Alchemy session
-    group_name : str
-        group_name in the database
-    kwargs : dict
-        keyword name and value
-    """
-    group = session.query(Groups).filter(Groups.name == group_name).first()
-
-    if group is not None:
-
-        group.keywords_query.delete()
-        for name, value in kwargs.items():
-            group.keywords.append(GroupKeywords(name=name, value=value))
-
-    else:
-
-        raise NoResultFound("group_name = '{}' was not found in database.".format(group_name))
-
-
-def update_group_keywords(session, group_name, **kwargs):
+def update_group_keywords(session, group_name, priority=None, **kwargs):
     """
     Function to update / add group keywords.
     Will create new group if group is not in database.
@@ -1055,8 +1108,11 @@ def update_group_keywords(session, group_name, **kwargs):
         group name in the database
     kwargs : dict
         keyword name and value
+    priority: int or None
+        use session priority if None
     """
-    # get group
+
+    priority = priority if priority is not None else session.info["SIMDB_PRIORITY"]
     group = session.query(Groups).filter(Groups.name == group_name).first()
 
     if group is not None:
@@ -1064,21 +1120,24 @@ def update_group_keywords(session, group_name, **kwargs):
         # update existing group keywords
         group_keywords = group.keywords_query.filter(GroupKeywords.name.in_(kwargs.keys())).all()
         for keyword in group_keywords:
-            keyword.value = kwargs.pop(keyword.name)
+            if priority >= keyword.priority:
+                keyword.value = kwargs.pop(keyword.name)
+                keyword.priority = priority
 
     else:
 
         # create new group
-        group = Groups(name=group_name)
+        group = Groups(name=group_name, priority=priority)
         session.add(group)
 
+    # add missing keywords
     for name, value in kwargs.items():
-        group.keywords.append(GroupKeywords(name=name, value=value))
+        group.keywords.append(GroupKeywords(name=name, value=value, priority=priority))
 
 
-def delete_all_group_keywords(session, group_name):
+def delete_group_keywords(session, group_name, group_keywords, priority=None):
     """
-    Function to delete all keywords for a group.
+    Function to delete a list of keyword of a group.
 
     Parameters
     ----------
@@ -1086,30 +1145,71 @@ def delete_all_group_keywords(session, group_name):
         SQL Alchemy session
     group_name : str
         group name in the database
+    group_keywords : list(str)
+        keyword names of group
+    priority: int or None
+        use session priority if None
     """
+
+    priority = priority if priority is not None else session.info["SIMDB_PRIORITY"]
+    group_keyword = session.query(GroupKeywords).filter(Groups.name == group_name) \
+        .filter(GroupKeywords.name.in_(group_keywords)).all()
+
+    for keyword in group_keyword:
+        if priority >= keyword.priority:
+            session.delete(keyword)
+
+
+def delete_all_group_keywords(session, group_name, priority=None):
+    """
+    Function to delete all keywords of a group.
+
+    Parameters
+    ----------
+    session : sqlalchemy.orm.session.Session
+        SQL Alchemy session
+    group_name : str
+        group name in the database
+    priority: int or None
+        use session priority if None
+    """
+
+    priority = priority if priority is not None else session.info["SIMDB_PRIORITY"]
     group = session.query(Groups).filter(Groups.name == group_name).first()
 
     if group is not None:
 
-        group.keywords_query.delete()
+        for keyword in group.keywords:
+            if priority >= keyword.priority:
+                session.delete(keyword)
 
     else:
 
         raise NoResultFound("group_name = '{}' was not found in database.".format(group_name))
 
 
+def set_group_keywords(session, group_name, priority=None, **kwargs):
+    """
+    Function to set group keywords. Overwrites the old!
+
+    Parameters
+    ----------
+    session : sqlalchemy.orm.session.Session
+        SQL Alchemy session
+    group_name : str
+        group_name in the database
+    kwargs : dict
+        keyword name and value
+    priority: int or None
+        use session priority if None
+    """
+
+    priority = priority if priority is not None else session.info["SIMDB_PRIORITY"]
+    delete_all_group_keywords(session, group_name, priority=priority)
+    update_group_keywords(session, group_name, priority=priority, **kwargs)
 
 
-def selectByKeyword(table, name, value):
-    '''Get mask for selection of entries by keyword.'''
-    return table[name] == value
 
-
-def selectByTag(table, tag):
-    '''Get mask for selection of entries by tag.'''
-    split = table.tags.str.split(",")
-    mask = [True if np.any(np.array(i[1]) == tag) else False for i in split.items()]
-    return pd.Series(mask, index=range(1, len(mask) + 1))
 
 def store_dict(entry_id,
                path,
@@ -1217,3 +1317,14 @@ def store_dict(entry_id,
     return sim
 
 
+# DEPRICATED
+# def selectByKeyword(table, name, value):
+#     '''Get mask for selection of entries by keyword.'''
+#     return table[name] == value
+#
+#
+# def selectByTag(table, tag):
+#     '''Get mask for selection of entries by tag.'''
+#     split = table.tags.str.split(",")
+#     mask = [True if np.any(np.array(i[1]) == tag) else False for i in split.items()]
+#     return pd.Series(mask, index=range(1, len(mask) + 1))
